Сделал прасинг настроек и их сохранение. Кнопка выбора маршрута

+++ app/src/main/java/ru/glorient/granitbk_n/avtoinformer/Stop.kt
@@ -2,13 +2,13 @@ package ru.glorient.granitbk_n.avtoinformer
 
 sealed class Stop {
     data class DefaultStop(
-        val id: Long,
+        val id: Int,
         val name: String,
         val time: String
     ) : Stop()
 
     data class NextStop(
-        val id: Long,
+        val id: Int,
         val name: String,
         val time: String
     ) : Stop()
